+++ lib/mutant/mutator/node/arguments.rb
@@ -25,7 +25,7 @@ module Mutant
           children.each_with_index do |removed, index|
             new_arguments = children.dup
             new_arguments.delete_at(index)
-            unless n_forward_arg?(removed) || removed_block_arg?(new_arguments) || only_mlhs?(new_arguments)
+            unless forward_type?(removed) || removed_block_arg?(new_arguments) || only_mlhs?(new_arguments)
               emit_type(*new_arguments)
             end
           end
@@ -36,7 +36,12 @@ module Mutant
         end
 
         def forward_arg?
-          children.last && n_forward_arg?(children.last)
+          children.any?(&method(:forward_type?))
+        end
+
+        def forward_type?(node)
+          n_forward_arg?(node) ||
+              ((n_restarg?(node) || n_kwrestarg?(node)) && node.children.empty?)
         end
 
         def removed_block_arg?(new_arguments)